wip: booking service in progress --> Be1,2 + be3

+++ parking-backend/src/main/java/com/swp391/parking/repository/VehicleRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 @Repository
 public interface VehicleRepository extends JpaRepository<Vehicle, Long> {
 
-    List<Vehicle> findByUserIdAndIsActiveTrue(Integer userId);
+    List<Vehicle> findByUserIdAndIsActiveTrue(Long userId);
 
     // Staff dùng khi xe vào/ra cổng
     Optional<Vehicle> findByLicensePlate(String licensePlate);
